fix prime

+++ brain_games/games/prime.py
@@ -7,7 +7,7 @@ DESCRIPTION = 'Answer "yes" if given number is prime. Otherwise answer "no".'
 
 def is_prime(num):
     if num > 1:
-        for i in range(2, num / 2):
+        for i in range(2, int(num / 2)):
             if num % i == 0:
                 return False
         return True
